:zap: Moved types to types/
