+++ portfolio/schema/skills.ts
@@ -1,6 +1,6 @@
 import mongoose from "mongoose"
 import { portfolioDb } from "../db/portfoliodb"
-import { SkillSchemaType } from "../utils/types"
+import { SkillSchemaType } from "../types/schema"
 
 const { Schema } = mongoose
