+++ portfolio/schema/token.ts
@@ -1,6 +1,6 @@
 import mongoose from "mongoose"
 import { portfolioDb } from "../db/portfoliodb"
-import { TokenSchemaType } from "../utils/types"
+import { TokenSchemaType } from "../types/schema"
 
 const { Schema } = mongoose
