+++ portfolio/utils/generateSecret.ts
@@ -1,5 +1,6 @@
 import crypto from "crypto"
 
+//? Generate a token secret
 const generateSecretToken = () => {
 	return crypto.randomBytes(64).toString("hex")
 }
