+++ portfolio/utils/handlePassword.ts
@@ -1,9 +1,11 @@
 import bcrypt from "bcrypt"
 
+//? Hash password
 const hashPassword = async (password: string) => {
 	return await bcrypt.hash(password, 10)
 }
 
+//? Compare passwords
 const comparePassword = async (
 	inputPassword: string,
 	hashedPassword: string
