+++ portfolio/utils/handleToken.ts
@@ -1,14 +1,16 @@
 import jwt from "jsonwebtoken"
-import { PayloadSchemaType } from "./types"
+import { PayloadSchemaType } from "../types/middleware"
 import {
 	PORTFOLIO_ACCESS_TOKEN_SECRET,
-	PORTFOLIO_REFRESH_TOKEN_SECRET
+	PORTFOLIO_REFRESH_TOKEN_SECRET,
+	PORTFOLIO_ACCESS_TOKEN_EXPIRATION_TIME
 } from "./constants"
+import { TOKEN } from "./constants"
 
 //? Generate accessToken
 const generateAccessToken = (payload: PayloadSchemaType) => {
 	return jwt.sign(payload, PORTFOLIO_ACCESS_TOKEN_SECRET, {
-		expiresIn: "45m"
+		expiresIn: PORTFOLIO_ACCESS_TOKEN_EXPIRATION_TIME
 	})
 }
 
@@ -17,11 +19,11 @@ const generateRefreshToken = (payload: PayloadSchemaType) => {
 	return jwt.sign(payload, PORTFOLIO_REFRESH_TOKEN_SECRET)
 }
 
-//? Verify token
+//? Verify token signature
 const verifyToken = (token: string, tokenType: string) => {
-	if (tokenType === "refresh_token")
+	if (tokenType === TOKEN.refreshToken)
 		return jwt.verify(token, PORTFOLIO_REFRESH_TOKEN_SECRET)
-	if (tokenType === "access_token")
+	if (tokenType === TOKEN.accessToken)
 		return jwt.verify(token, PORTFOLIO_ACCESS_TOKEN_SECRET)
 }
